SQ additional test coverage & fixes
